requirements.txt
google-generativeai==0.5.4
google-cloud-texttospeech==2.16.3
google-cloud-speech==2.26.0
numpy==2.0.0
pandas==2.2.2
uvicorn==0.30.1
fastapi==0.111.0
joblib==1.4.2